Utility function to process API response
Wrote a utility function to create an array of Teams from a SportsRadar API standings response. The response is a deeply nested object, so this method will transform the response into an array and remove unwanted data points. An array will be easier to work with when displaying this data with a ReactNative FlatList