SeasonalStats screen drafting
This screen shows detailed information about a team's stats in the given season. Press the TouchableOpacity 'toggle hitting' updates local state to showHitting: true, which will be used to conditionally render the hitting stats for the team's season. The other categories from the API, pitching and fielding, are set to follow