Brought in Redux, calling API with Thunk action
Integrated Redux into the project, and am currently fetching live data from the SportRadar API for the SeasonalStats screen. Need to settle on a solution for structuring the data retrieved from the API and the proper way to hold it in the Redux Store. The data given back is much more than I need for the context of this application, so I will use some helper util functions to strip it down for storing in Redux as it is requested from the API