Pull in real data from API for each team
Passing team info through navigation props to immediately render the team's name, year for the stats, and the team's logo. On Component Mount, a thunk action is dispatched to fetch the stats from the SportRadar API, which will display in the StatsTable at the bottom once the data is loaded in successfully. Set a maxHeight for the Team Logo because the images I am using are different sizes.